refactor(test): Split copyToChildren tests into separate files per group
Split monolithic CopyToChildrenClasses.kt and CopyToChildrenEdgeCaseClasses.kt
into individual files following the combineMapping pattern.

Changes:
- Split into: Basic, ObjectTarget, Nested, ComplexTypes, Visibility
- Update test files with corresponding splits
- Remove edgeCase subdirectory
- Each test group now has its own dedicated file

This makes tests easier to navigate and maintain.
