+++ test/src/commonTest/kotlin/me/tbsten/cream/test/copyToChildren/ObjectTargetTest.kt
@@ -1,0 +1,14 @@
+package me.tbsten.cream.test.copyToChildren
+
+import kotlin.test.Test
+import kotlin.test.assertEquals
+
+class ObjectTargetTest {
+    @Test
+    fun dataObjectToDataObject() {
+        val source = EmptyChild1
+        val target = source.copyToEmptyChild2()
+
+        assertEquals(EmptyChild2, target)
+    }
+}
